+++ packages/core/tests/create-field.spec.ts
@@ -1,10 +1,8 @@
-import matchers from '@testing-library/jest-dom/matchers';
+import '@testing-library/jest-dom/vitest';
 import { expect, describe, test, vi, beforeEach, afterEach } from 'vitest';
 import { waitFor, screen } from '@testing-library/dom';
 import { createInputElement, createDOM, cleanupDOM } from './common';
 import { createField } from '../src';
-
-expect.extend(matchers);
 
 function createContentEditable() {
   const input = document.createElement('div');
@@ -45,7 +43,7 @@ describe('Custom controls with createField', () => {
     formElement.appendChild(hiddenElement);
 
     expect(formElement.querySelectorAll('input[name="test"]').length).to.equal(
-      1
+      1,
     );
 
     const { field } = createField({ name: 'test' });
@@ -53,7 +51,7 @@ describe('Custom controls with createField', () => {
     field(inputElement);
 
     expect(formElement.querySelectorAll('input[name="test"]').length).to.equal(
-      1
+      1,
     );
   });
 
@@ -70,7 +68,7 @@ describe('Custom controls with createField', () => {
 
     await waitFor(() => {
       expect(
-        formElement.querySelectorAll('input[name="test"]').length
+        formElement.querySelectorAll('input[name="test"]').length,
       ).to.equal(1);
       expect(formElement.querySelector('input[name="test"]')).toBeVisible();
     });
@@ -100,7 +98,7 @@ describe('Custom controls with createField', () => {
 
     await waitFor(() => {
       const hiddenElement = document.querySelector(
-        'input[name="test"]'
+        'input[name="test"]',
       ) as HTMLInputElement;
 
       expect(hiddenElement).not.to.be.null;
@@ -110,14 +108,14 @@ describe('Custom controls with createField', () => {
 
     await waitFor(() => {
       const hiddenElement = document.querySelector(
-        'input[name="test"]'
+        'input[name="test"]',
       ) as HTMLInputElement;
 
       expect(hiddenElement.value).to.equal('new value');
       expect(inputListener).toHaveBeenCalledWith(
         expect.objectContaining({
           target: hiddenElement,
-        })
+        }),
       );
       expect(blurListener).not.toHaveBeenCalled();
 
@@ -126,7 +124,7 @@ describe('Custom controls with createField', () => {
       expect(blurListener).toHaveBeenCalledWith(
         expect.objectContaining({
           target: hiddenElement,
-        })
+        }),
       );
     });
 
@@ -153,7 +151,7 @@ describe('Custom controls with createField', () => {
 
     await waitFor(() => {
       const hiddenElement = document.querySelector(
-        'input[name="test"]'
+        'input[name="test"]',
       ) as HTMLInputElement;
 
       expect(hiddenElement).not.to.be.null;
@@ -195,7 +193,7 @@ describe('Custom controls with createField', () => {
     await waitFor(() => {
       expect(inputElement).toHaveAttribute(
         'data-felte-validation-message',
-        'a message'
+        'a message',
       );
     });
     hiddenElement.removeAttribute('data-felte-validation-message');
@@ -235,7 +233,7 @@ describe('Custom controls with createField', () => {
       input.type = 'text';
       return input;
     }
-    
+
     const formElement = screen.getByRole('form') as HTMLFormElement;
     const inputElement = createTextInput();
     inputElement.value = 'old value';
@@ -248,7 +246,7 @@ describe('Custom controls with createField', () => {
 
     onChange('new value');
     expect(inputElement.value).to.equal('new value');
-    
+
     onBlur();
     expect(inputElement.value).to.equal('new value');
   });