+++ packages/reporter-svelte/src/ValidationMessage.svelte
@@ -1,6 +1,5 @@
 <script>
   import { onMount } from 'svelte';
-  import { writable } from 'svelte/store';
   import { _get, getPath } from '@felte/common';
   import { errorStores, warningStores } from './stores';
 
@@ -17,17 +16,18 @@
   }
 
   onMount(() => {
-    const path = errorFor;
-    errorPath = getPath(element, path);
-    const formElement = getFormElement();
-    if (!formElement) errors = writable({});
-    else if (level === 'error') errors = errorStores[formElement.dataset.felteReporterSvelteId];
-    else errors = warningStores[formElement.dataset.felteReporterSvelteId];
+    setTimeout(() => {
+      const path = errorFor;
+      errorPath = getPath(element, path);
+      const formElement = getFormElement();
+      if (level === 'error') errors = errorStores[formElement.dataset.felteReporterSvelteId];
+      else errors = warningStores[formElement.dataset.felteReporterSvelteId];
+    })
   });
   $: messages = errorPath && _get($errors, errorPath)
 </script>
 
-<div bind:this={element} style="display: none;" />
+<div bind:this={element} style="display: none;"></div>
 {#if !$$slots.placeholder || messages}
   <slot {messages}></slot>
 {:else}