+++ packages/felte/CHANGELOG.md
@@ -1,5 +1,17 @@
 # felte
 
+## 1.3.0
+
+### Minor Changes
+
+- 7c3dffc: Update peer dependencies to support Svelte 5
+
+### Patch Changes
+
+- 7c3dffc: Fix typescript issues
+- Updated dependencies [7c3dffc]
+  - @felte/core@1.4.4
+
 ## 1.2.14
 
 ### Patch Changes
@@ -205,7 +217,7 @@
   If you have an extender using `addWarnValidator`, you must update it by calling `addValidator` instead with the following options:
 
   ```javascript
-  addValidator(yourValidationFunction, { level: 'warning' });
+  addValidator(yourValidationFunction, { level: "warning" });
   ```
 
 - b007b7f: BREAKING: Helpers have been completely reworked.
@@ -499,7 +511,7 @@
   If you have an extender using `addWarnValidator`, you must update it by calling `addValidator` instead with the following options:
 
   ```javascript
-  addValidator(yourValidationFunction, { level: 'warning' });
+  addValidator(yourValidationFunction, { level: "warning" });
   ```
 
 ### Minor Changes