+++ packages/preact/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @felte/preact
 
+## 1.2.14
+
+### Patch Changes
+
+- 7c3dffc: Fix typescript issues
+- Updated dependencies [7c3dffc]
+  - @felte/core@1.4.4
+
 ## 1.2.13
 
 ### Patch Changes