+++ packages/validator-yup/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @felte/validator-yup
 
+## 1.1.4
+
+### Patch Changes
+
+- 7c3dffc: Fix typescript issues
+- Updated dependencies [7c3dffc]
+  - @felte/common@1.1.9
+
 ## 1.1.3
 
 ### Patch Changes
@@ -126,7 +134,7 @@
   If you have an extender using `addWarnValidator`, you must update it by calling `addValidator` instead with the following options:
 
   ```javascript
-  addValidator(yourValidationFunction, { level: 'warning' });
+  addValidator(yourValidationFunction, { level: "warning" });
   ```
 
 - b007b7f: Pass a new property `stage` to extenders to distinguish between setup, mount and update stages
@@ -271,7 +279,7 @@
   If you have an extender using `addWarnValidator`, you must update it by calling `addValidator` instead with the following options:
 
   ```javascript
-  addValidator(yourValidationFunction, { level: 'warning' });
+  addValidator(yourValidationFunction, { level: "warning" });
   ```
 
 ### Patch Changes